+++ api/src/Blocks.ts
@@ -4,13 +4,13 @@ import { Compact, GenericExtrinsic, Vec, u64 } from '@polkadot/types';
 import { isHex, isNumber, isU8a } from '@polkadot/util';
 import { FrameSystemEventRecord } from '@polkadot/types/lookup';
 import { HeaderExtended } from '@polkadot/api-derive/types';
+import { HexString } from '@polkadot/util/types';
 import { Observable } from 'rxjs';
 import { PromiseResult } from '@polkadot/api/types';
 
 import { CreateType } from './create-type';
 import { GearApi } from './GearApi';
 import { GetBlockError } from './errors';
-import { Hex } from './types';
 
 export class GearBlock {
   subscribeNewHeads: PromiseResult<() => Observable<HeaderExtended>>;
@@ -24,7 +24,7 @@ export class GearBlock {
    * @param hash
    * @returns
    */
-  async get(hash: Hex | Uint8Array): Promise<SignedBlock>;
+  async get(hash: HexString | Uint8Array): Promise<SignedBlock>;
 
   /**
    * Get data of particular block by blockNumber
@@ -38,14 +38,14 @@ export class GearBlock {
    * @param hashOrNumber
    * @returns
    */
-  async get(hashOrNumber: Hex | Uint8Array | number): Promise<SignedBlock>;
+  async get(hashOrNumber: HexString | Uint8Array | number): Promise<SignedBlock>;
 
   /**
    * Get data of particular block by blockNumber or blockHash
    * @param hashOrNumber
    * @returns
    */
-  async get(hashOrNumber: Hex | Uint8Array | number): Promise<SignedBlock> {
+  async get(hashOrNumber: HexString | Uint8Array | number): Promise<SignedBlock> {
     const hash = isU8a(hashOrNumber) || isHex(hashOrNumber) ? hashOrNumber : await this.getBlockHash(+hashOrNumber);
     try {
       return await this.api.rpc.chain.getBlock(hash);
@@ -83,7 +83,7 @@ export class GearBlock {
    * ### Get block's timestamp by blockHash
    * @param hash
    */
-  async getBlockTimestamp(hash: Hex | Uint8Array): Promise<Compact<u64>>;
+  async getBlockTimestamp(hash: HexString | Uint8Array): Promise<Compact<u64>>;
 
   /**
    * ### Get block's timestamp by blockNumber
@@ -91,7 +91,7 @@ export class GearBlock {
    */
   async getBlockTimestamp(number: number): Promise<Compact<u64>>;
 
-  async getBlockTimestamp(blockOrHashOrNumber: Hex | Uint8Array | number | SignedBlock): Promise<Compact<u64>> {
+  async getBlockTimestamp(blockOrHashOrNumber: HexString | Uint8Array | number | SignedBlock): Promise<Compact<u64>> {
     const block =
       isHex(blockOrHashOrNumber) || isU8a(blockOrHashOrNumber) || isNumber(blockOrHashOrNumber)
         ? await this.get(blockOrHashOrNumber)