+++ api/src/Code.ts
@@ -1,8 +1,9 @@
 import { Bytes, Option } from '@polkadot/types';
+import { HexString } from '@polkadot/util/types';
 import { ISubmittableResult } from '@polkadot/types/types';
 import { SubmittableExtrinsic } from '@polkadot/api/types';
 
-import { CodeMetadata, CodeStorage, Hex } from './types';
+import { CodeMetadata, CodeStorage } from './types';
 import { generateCodeHash, getIdsFromKeys, validateCodeId } from './utils';
 import { GearTransaction } from './Transaction';
 
@@ -14,7 +15,7 @@ export class GearCode extends GearTransaction {
    */
   async upload(
     code: Buffer | Uint8Array,
-  ): Promise<{ codeHash: Hex; submitted: SubmittableExtrinsic<'promise', ISubmittableResult> }> {
+  ): Promise<{ codeHash: HexString; submitted: SubmittableExtrinsic<'promise', ISubmittableResult> }> {
     const codeHash = generateCodeHash(code);
     await validateCodeId(codeHash, this._api);
 
@@ -36,7 +37,7 @@ export class GearCode extends GearTransaction {
    * ### Get code storage
    * @param codeId
    */
-  async storage(codeId: Hex): Promise<CodeStorage> {
+  async storage(codeId: HexString): Promise<CodeStorage> {
     return this._api.query.gearProgram.codeStorage(codeId) as unknown as CodeStorage;
   }
 
@@ -44,7 +45,7 @@ export class GearCode extends GearTransaction {
    * ### Get static pages of code
    * @param codeId
    */
-  async staticPages(codeId: Hex): Promise<number | null> {
+  async staticPages(codeId: HexString): Promise<number | null> {
     const storage = await this.storage(codeId);
     return storage.isSome ? storage.unwrap().staticPages.toNumber() : null;
   }
@@ -53,7 +54,7 @@ export class GearCode extends GearTransaction {
    * ### Get all ids of codes uploaded on connected chain
    * @returns array of code ids uploaded on chain
    */
-  async all(): Promise<Hex[]> {
+  async all(): Promise<HexString[]> {
     const prefix = this._api.query.gearProgram.metadataStorage.keyPrefix();
     const keys = await this._api.rpc.state.getKeys(prefix);
     const codeIds = getIdsFromKeys(keys, prefix);