+++ api/src/types/common.ts
@@ -1,8 +1,6 @@
 import { u128, u64 } from '@polkadot/types';
 import { AnyNumber } from '@polkadot/types/types';
 
-export type Hex = `0x${string}`;
-
 export type Value = AnyNumber | u128;
 
 export type GasLimit = AnyNumber | u64;