+++ api/src/utils/address.ts
@@ -1,9 +1,8 @@
+import { HexString } from '@polkadot/util/types';
 import { Keyring } from '@polkadot/api';
 import { u8aToHex } from '@polkadot/util';
 
-import { Hex } from '../types';
-
-export function decodeAddress(publicKey: string): Hex {
+export function decodeAddress(publicKey: string): HexString {
   return u8aToHex(new Keyring().decodeAddress(publicKey));
 }